+++ examples/node_based_editor/lib/editor.dart
@@ -49,7 +49,7 @@ class _EditorState extends State<Editor> {
             IconButton(
               icon: const Icon(Icons.zoom_in),
               tooltip: 'Zoom In',
-              onPressed: () => matrix.forceUpdate(matrix.value
+              onPressed: () => matrix.set(force: true, matrix.value
                 ..translate(size.width / 2, size.height / 2)
                 ..scale(1.1, 1.1)
                 ..translate(-(size.width / 2), -(size.height / 2))),
@@ -57,7 +57,7 @@ class _EditorState extends State<Editor> {
             IconButton(
               icon: const Icon(Icons.zoom_out),
               tooltip: 'Zoom Out',
-              onPressed: () => matrix.forceUpdate(matrix.value
+              onPressed: () => matrix.set(force: true, matrix.value
                 ..translate(size.width / 2, size.height / 2)
                 ..scale(0.9, 0.9)
                 ..translate(-(size.width / 2), -(size.height / 2))),
@@ -103,7 +103,7 @@ class _EditorState extends State<Editor> {
                 child: SizedBox.expand(
                   child: GestureDetector(
                     onPanUpdate: (details) {
-                      matrix.forceUpdate(matrix.value
+                      matrix.set(force: true, matrix.value
                         ..translate(details.delta.dx, details.delta.dy));
                     },
                     child: CustomPaint(
@@ -442,7 +442,7 @@ class _EditorState extends State<Editor> {
                         color: () {
                           final colors = Theme.of(context).colorScheme;
                           final color =
-                              dragAccept() ? colors.primary : colors.background;
+                              dragAccept() ? colors.primary : colors.surface;
                           return color.withOpacity(0.1);
                         }(),
                       ),