+++ packages/signals_core/lib/src/core/signal.dart
@@ -139,7 +139,7 @@ part of 'signals.dart';
 /// }
 ///
 /// class _CounterWidgetState extends State<CounterWidget> with SignalsAutoDisposeMixin {
-///   late final counter = createSignal(this, 0);
+///   late final counter = createSignal(context, 0);
 ///
 ///   @override
 ///   Widget build(BuildContext context) {
@@ -205,7 +205,207 @@ part of 'signals.dart';
 /// {@endtemplate}
 class Signal<T> extends ReadonlySignal<T> {
   /// {@template signal}
-  /// ...
+  /// The `signal` function creates a new signal. A signal is a container for a value that can change over time. You can read a signal's value or subscribe to value updates by accessing its `.value` property.
+  ///
+  /// ```dart
+  /// import 'package:signals/signals.dart';
+  ///
+  /// final counter = signal(0);
+  ///
+  /// // Read value from signal, logs: 0
+  /// print(counter.value);
+  ///
+  /// // Write to a signal
+  /// counter.value = 1;
+  /// ```
+  ///
+  /// Signals can be created globally, inside classes or functions. It's up to you how you want to structure your app.
+  ///
+  /// It is not recommended to create signals inside effects or computed, as this will create a new signal every time the effect or computed is triggered. This can lead to unexpected behavior.
+  ///
+  /// In Flutter do not create signals inside `build` methods, as this will create a new signal every time the widget is rebuilt.
+  ///
+  /// ## Writing to a signal
+  ///
+  /// Writing to a signal is done by setting its `.value` property. Changing a signal's value synchronously updates every [computed](/core/computed) and [effect](/core/effect) that depends on that signal, ensuring your app state is always consistent.
+  ///
+  /// ## .peek()
+  ///
+  /// In the rare instance that you have an effect that should write to another signal based on the previous value, but you _don't_ want the effect to be subscribed to that signal, you can read a signals's previous value via `signal.peek()`.
+  ///
+  /// ```dart
+  /// final counter = signal(0);
+  /// final effectCount = signal(0);
+  ///
+  /// effect(() {
+  /// 	print(counter.value);
+  ///
+  /// 	// Whenever this effect is triggered, increase `effectCount`.
+  /// 	// But we don't want this signal to react to `effectCount`
+  /// 	effectCount.value = effectCount.peek() + 1;
+  /// });
+  /// ```
+  ///
+  /// Note that you should only use `signal.peek()` if you really need it. Reading a signal's value via `signal.value` is the preferred way in most scenarios.
+  ///
+  /// ## .value
+  ///
+  /// The `.value` property of a signal is used to read or write to the signal. If used inside an effect or computed, it will subscribe to the signal and trigger the effect or computed whenever the signal's value changes.
+  ///
+  /// ```dart
+  /// final counter = signal(0);
+  ///
+  /// effect(() {
+  /// 	print(counter.value);
+  /// });
+  ///
+  /// counter.value = 1;
+  /// ```
+  ///
+  /// ## .previousValue
+  ///
+  /// The `.previousValue` property of a signal is used to read the previous value of the signal. If used inside an effect or computed, it will not subscribe to the signal and not trigger the effect or computed whenever the signal's value changes.
+  ///
+  /// ```dart
+  /// final counter = signal(0);
+  ///
+  /// effect(() {
+  /// 	print('Current value: ${counter.value}');
+  /// 	print('Previous value: ${counter.previousValue}');
+  /// });
+  ///
+  /// counter.value = 1;
+  /// ```
+  ///
+  /// ## Force Update
+  ///
+  /// If you want to force an update for a signal, you can call the `.set(..., force: true)` method. This will trigger all effects and mark all computed as dirty.
+  ///
+  /// ```dart
+  /// final counter = signal(0);
+  /// counter.set(1, force: true);
+  /// ```
+  ///
+  /// ## Disposing
+  ///
+  /// ### Auto Dispose
+  ///
+  /// If a signal is created with autoDispose set to true, it will automatically dispose itself when there are no more listeners.
+  ///
+  /// ```dart
+  /// final s = signal(0, autoDispose: true);
+  /// s.onDispose(() => print('Signal destroyed'));
+  /// final dispose = s.subscribe((_) {});
+  /// dispose();
+  /// final value = s.value; // 0
+  /// // prints: Signal destroyed
+  /// ```
+  ///
+  /// A auto disposing signal does not require its dependencies to be auto disposing. When it is disposed it will freeze its value and stop tracking its dependencies.
+  ///
+  /// ```dart
+  /// final s = signal(0);
+  /// s.dispose();
+  /// final c = computed(() => s.value);
+  /// // c will not react to changes in s
+  /// ```
+  ///
+  /// You can check if a signal is disposed by calling the `.disposed` method.
+  ///
+  /// ```dart
+  /// final s = signal(0);
+  /// print(s.disposed); // false
+  /// s.dispose();
+  /// print(s.disposed); // true
+  /// ```
+  ///
+  /// ### On Dispose Callback
+  ///
+  /// You can attach a callback to a signal that will be called when the signal is destroyed.
+  ///
+  /// ```dart
+  /// final s = signal(0);
+  /// s.onDispose(() => print('Signal destroyed'));
+  /// s.dispose();
+  /// ```
+  ///
+  /// ## Flutter
+  ///
+  /// In Flutter if you want to create a signal that automatically disposes itself when the widget is removed from the widget tree and rebuilds the widget when the signal changes, you can use the `createSignal` inside a stateful widget.
+  ///
+  /// ```dart
+  /// import 'package:flutter/material.dart';
+  /// import 'package:signals/signals_flutter.dart';
+  ///
+  /// class CounterWidget extends StatefulWidget {
+  ///   @override
+  ///   _CounterWidgetState createState() => _CounterWidgetState();
+  /// }
+  ///
+  /// class _CounterWidgetState extends State<CounterWidget> with SignalsAutoDisposeMixin {
+  ///   late final counter = createSignal(context, 0);
+  ///
+  ///   @override
+  ///   Widget build(BuildContext context) {
+  ///     return Scaffold(
+  ///       body: Center(
+  ///         child: Column(
+  ///           mainAxisAlignment: MainAxisAlignment.center,
+  ///           children: [
+  ///             Text('Counter: $counter'),
+  ///             ElevatedButton(
+  ///               onPressed: () => counter.value++,
+  ///               child: Text('Increment'),
+  ///             ),
+  ///           ],
+  ///         ),
+  ///       ),
+  ///     );
+  ///   }
+  /// }
+  /// ```
+  ///
+  /// No `Watch` widget or extension is needed, the signal will automatically dispose itself when the widget is removed from the widget tree.
+  ///
+  /// The `SignalsAutoDisposeMixin` is a mixin that automatically disposes all signals created in the state when the widget is removed from the widget tree.
+  ///
+  /// ## Testing
+  ///
+  /// Testing signals is possible by converting a signal to a stream and testing it like any other stream in Dart.
+  ///
+  /// ```dart
+  /// test('test as stream', () {
+  ///   final s = signal(0);
+  ///   final stream = s.toStream(); // create a stream of values
+  ///
+  ///   s.value = 1;
+  ///   s.value = 2;
+  ///   s.value = 3;
+  ///
+  ///   expect(stream, emitsInOrder([0, 1, 2, 3]));
+  /// });
+  /// ```
+  ///
+  /// `emitsInOrder` is a matcher that will check if the stream emits the values in the correct order which in this case is each value after a signal is updated.
+  ///
+  /// You can also override the initial value of a signal when testing. This is is useful for mocking and testing specific value implementations.
+  ///
+  /// ```dart
+  /// test('test with override', () {
+  ///   final s = signal(0).overrideWith(-1);
+  ///
+  ///   final stream = s.toStream();
+  ///
+  ///   s.value = 1;
+  ///   s.value = 2;
+  ///   s.value = 3;
+  ///
+  ///   expect(stream, emitsInOrder([-1, 1, 2, 3]));
+  /// });
+  /// ```
+  ///
+  /// `overrideWith` returns a new signal with the same global id sets the value as if it was created with it. This can be useful when using async signals or global signals used for dependency injection.
+  /// @link https://dartsignals.dev/core/signal
   /// {@endtemplate}
   Signal(
     T val, {
@@ -242,7 +442,7 @@ class Signal<T> extends ReadonlySignal<T> {
 
   /// Check if the signal is lazy and has not had a value set
   @override
-  bool get isLazy => _lazy == true;
+  bool get isInitialized => _lazy == false;
   bool _lazy = false;
 
   @override
@@ -255,7 +455,7 @@ class Signal<T> extends ReadonlySignal<T> {
   /// Force update a value
   @Deprecated('Use .set(..., force: true) instead')
   void forceUpdate([T? val]) {
-    assert(!_lazy && val != null, 'Lazy signal must be initilized first');
+    assert(!_lazy && val != null, 'Lazy signal must be initialized first');
     this.set(val ?? value, force: true);
   }
 
@@ -272,11 +472,23 @@ class Signal<T> extends ReadonlySignal<T> {
     }
   }
 
+  /// Optional method to check if to values are the same
+  bool Function(T a, T b) equalityCheck = (T a, T b) => a == b;
+
   /// Update the current value.
   ///
   /// `force` an update if needed (if the update would
   /// not pass the == check)
-  void set(T val, {bool force = false}) {
+  bool set(T val, {bool force = false}) {
+    if (!_lazy && !(force || !equalityCheck(val, _value))) {
+      return false;
+    }
+
+    if (_callDepth > _maxCallDepth) {
+      // coverage:ignore-start
+      throw EffectCycleDetectionError();
+      // coverage:ignore-end
+    }
     if (_lazy) {
       _value = val;
       _initialValue = val;
@@ -290,21 +502,22 @@ class Signal<T> extends ReadonlySignal<T> {
       }
       _previousValue = _value ?? _initialValue;
       _value = val;
-      _version++;
-      _globalVersion++;
-
-      _startBatch();
-      try {
-        _notifyAllTargets();
-      } finally {
-        _endBatch();
-      }
-
-      assert(() {
-        SignalsObserver.instance?.onSignalUpdated(this, val);
-        return true;
-      }());
     }
+    _version++;
+    _globalVersion++;
+
+    _startBatch();
+    try {
+      _notifyAllTargets();
+    } finally {
+      _endBatch();
+    }
+
+    assert(() {
+      SignalsObserver.instance?.onSignalUpdated(this, val);
+      return true;
+    }());
+    return true;
   }
 
   /// Set the current value
@@ -530,7 +743,7 @@ class Signal<T> extends ReadonlySignal<T> {
 /// }
 ///
 /// class _CounterWidgetState extends State<CounterWidget> with SignalsAutoDisposeMixin {
-///   late final counter = createSignal(this, 0);
+///   late final counter = createSignal(context, 0);
 ///
 ///   @override
 ///   Widget build(BuildContext context) {