+++ packages/signals_flutter/lib/signals_flutter.dart
@@ -1,6 +1,6 @@
 library signals_flutter;
 
-export 'package:signals_core/signals_core.dart';
+export 'signals_core.dart';
 export 'src/extensions/signal.dart';
 export 'src/extensions/value_listenable.dart';
 export 'src/extensions/value_notifier.dart';
@@ -8,5 +8,6 @@ export 'src/core/signal_value_listenable.dart';
 export 'src/watch/watch.dart';
 export 'src/core/signal.dart';
 export 'src/core/computed.dart';
-export 'src/mixin.dart';
+export 'src/mixins/auto_dispose.dart';
 export 'src/flutter/ticker.dart';
+export 'src/mixins/signals.dart';