+++ packages/signals_flutter/lib/src/flutter/inherited_widget/readonly.dart
@@ -3,7 +3,7 @@ import 'package:flutter/widgets.dart';
 import '../../../signals_flutter.dart';
 import 'provider.dart';
 
-/// Convience method to pass signals with [InheritedWidget]
+/// Convenience method to pass signals with [InheritedWidget]
 class ReadonlySignalProvider<T>
     extends InheritedSignalProvider<T, ReadonlySignal<T>> {
   /// Create a new [ReadonlySignal] by value to provide in a widget tree
@@ -20,7 +20,7 @@ class ReadonlySignalProvider<T>
     required super.value,
   }) : super.value();
 
-  /// Look up a singal by its type
+  /// Look up a signal by its type
   static ReadonlySignal<T> of<T>(BuildContext context, {bool listen = true}) {
     return InheritedSignalProvider.of<ReadonlySignalProvider<T>>(
       context,